feat: moved signup into different files
