+++ frontend/src/components/LoginForm.tsx
@@ -69,6 +69,15 @@ export default function LoginForm() {
           {loginMutation.isPending ? 'Logging in...' : 'Login'}
         </button>
       </form>
+
+      <div className="mt-4 text-center text-sm">
+        <p className="text-gray-600 dark:text-gray-300">
+          Don't have an account?{' '}
+          <a href="/signup" className="hover:underline font-semibold">
+            Sign up
+          </a>
+        </p>
+      </div>
     </div>
   )
 }
